Added `de/structs` module
